requirements.txt
# LangChain for RAG pipeline
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# Vector database and embeddings
sentence-transformers>=2.2.2
elasticsearch>=8.11.0

# Document processing
unstructured[pdf]>=0.10.0  # Includes PDF processing dependencies
tiktoken>=0.5.0

# Web interface
streamlit>=1.28.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
